# aibots-api-develop/openapi/full.yaml
openapi: 3.0.3
info:
  title: AIBots Full API
  description: This document describes the Full API interface for AIBots
  version: 1.0.0
  contact:
    name: Moonshot Team
    url: 'https://www.tech.gov.sg/capability-centre-dsaid'
    email: david_tw_lee@tech.gov.sg
servers:
  - url: 'https://aibots.gov.sg/v{major}.{minor}'
    description: Full server address hosted on cloud
    variables:
      major:
        enum:
          - "1"
        default: "1"
      minor:
        enum:
          - "0"
        default: "0"
...
paths:
...
  '/sso/moonshot/logins':
    post:
      tags:
        - Login
      summary: Login via Moonshot SSO
      description: >-
        This API is part of the sutie of APIs that facilitate logins via SSO. 
        The endpoint is structured as such to allow different forms of SSO 
        integrations, with this being the primary form of integration for 
        all Moonshot products.
      operationId: login_moonshot_sso
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SSOMoonshotLoginPOST"
      responses:
        '201':
          description: Successfully logged a user in via Moonshot SSO.
          headers:
            Authorization:
              schema:
                type: string
                format: jwt
              example: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJEU0FJRCwgR292dGVjaCIsInN1YiI6IntcIm5hbWVcIjpcIkRhdmlkdHdsZWVcIixcImlkXCI6XCIwNGUwNTg5Ny0zYzYyLTUyYWEtODY0MC1jYTEyNjQ0MWU1MmNcIixcImVtYWlsXCI6XCJkYXZpZHR3bGVlQGRzYWlkLmdvdi5zZ1wiLFwibmV3X3VzZXJcIjpmYWxzZSxcInNjb3Blc1wiOltdfSIsImF1ZCI6IjA0ZTA1ODk3LTNjNjItNTJhYS04NjQwLWNhMTI2NDQxZTUyYyIsIm5iZiI6MTcwMzA2MDIwNSwiaWF0IjoxNzAzMDYwMjA1LCJleHAiOjE3MDMwNjc0MDUsImp0aSI6ImQxZDIzZTUwLTc3N2UtNDQ5OS1iNzFkLWQwZTE2ZGU2OTc4NSIsImF0X2hhc2giOiJKcDgtbGtCT3lzV1VNelBDbTNSNFNBIn0.i5HMBZHGuNVQggDSad8s1BlfYb7DUKzq1WRZalQ-xOc
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserLogin'
        '401':
          $ref: '#/components/responses/AuthenticationError'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
security:
  - AtlasAuth: []
components:
  schemas:
...
    APIError:
      description: Generic schema for expressing API Errors
      type: object
      properties:
        code:
          description: API Error code associated with the error
          type: integer
          format: int32
        message:
          description: Error message associated with the error
          type: string
        request:
          description: Request details associated with the error
          type: object
        details:
          description: Other details associated with the error
          type: object
      required:
        - code
        - message
...
    ScopeStr:
      description: Representation of a Scope string
      type: string
      pattern: >- 
        (?<product>(?<productName>[A-Za-z0-9_]+)-(?<semVer>[A-Za-z0-9]*\.?[A-Za-z0-9]*\.?[A-Za-z0-9]*.(a|alpha|b|beta|rc\.\d*)?):)?(?<resource>[A-Za-z0-9_]+(?<fields>[A-Za-z0-9_\.\*,])*(?<attributes>-[A-Za-z0-9]+)*)(?<accessLevel>:(create_one|create_many|create|read_one|read_many|read|update_one|update_many|update|delete_one|delete_many|delete|allow|deny|,)*)?
    ScopeList:
      description: List of Scope strings
      type: array
      items:
        $ref: '#/components/schemas/ScopeStr'
      default: []
...
    SSOMoonshotLoginPOST:
      description: >-
        Request parameters for logging in via Moonshot infrastructure SSO
      type: object
      properties:
        accessToken:
          description: AWS SSO Access Token
          type: string
        identity:
          description: AWS SSO identity
          type: string
        data:
          description: AWS SSO data
          type: string
      required:
        - accessToken
        - identity
        - data
...
    UserLogin:
      description: Login details of a User
      type: object
      properties:
        id:
          description: ID of the User
          type: string
          format: uuid
        name:
          description: Name of the User
          type: string
          maxLength: 200
        email:
          description: Email of the User
          type: string
          format: email
        new_user:
          description: Indicates that this is the first time a user is logging in
          type: boolean
          default: false
        token:
          description: JWT token value issued to a user
          type: string
          format: jwt
        refreshToken:
          description: Refresh token value issued to a user
          type: string
        refreshCount:
          description: Number of refresh attempts
          type: integer
          minimum: 0
        maxRefresh:
          description: Maximum number of allowable refresh attempts
          type: integer
          minimum: 0
        timestamp:
          description: Timestamp when the Login was generated
          type: string
          format: date-time
        expiry:
          description: Expiry timestamp for the token
          type: string
          format: date-time
        scopes:
          $ref: '#/components/schemas/ScopeList'
        role:
          description: ID of the associated User's Role
          type: string
          format: uuid
        superuser:
          description: Indicates if the User is a superuser
          type: boolean
          default: false
        agency:
          description: Agency code details of the User
          type: string
          nullable: true
          default: null
        agencyId:
          description: Agency ID details of the User
          type: string
          format: uuid
          nullable: true
          default: null
      required:
        - id
        - name
        - email
        - new_user
        - timestamp
        - expiry
...
  responses:
...
    AuthenticationError:
      description: Authentication error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/APIError'
          example:
            code: 401
            message: Unauthorised usage of the API
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/APIError'
          example:
            code: 500
            message: Internal server error

    # SCIM responses
...
  securitySchemes:
    AtlasAuth:
      description: Atlas JWT authentication token
      type: http
      scheme: bearer
      bearerFormat: JWT
    AtlasAPIKey:
      description: Atlas API Key authentication
      type: apiKey
      in: header
      name: X-ATLAS-Key